% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimaticInfo.R, R/StreamflowEnergy.R
\docType{data}
\name{ClimaticInfo}
\alias{ClimaticInfo}
\title{Climatic indicators for forecasting}
\format{
A four column data.frame with
\describe{
\item{Date}{It is the Date for the observed values}
\item{U1}{The Zonal Winds values}
\item{SST2}{The sea surface temperature in part of the Atlancic}
\item{NINO3}{The sea surface temperature at the Nino 3.4 region in the Pacific}
}

A five column data.frame with
\describe{
\item{Date}{It is the Date for the observed values}
\item{Subsystem_N}{orthern Brazilian Subsystem}
\item{Subsystem_NE}{North East Brazilian Subsystem}
\item{Subsystem_S}{Southern Brazilian Subsystem}
\item{Subsystem_SE}{South East / Mid West Brazilian Subsystem}
}
}
\usage{
ClimaticInfo

ClimaticInfo
}
\description{
Three climatic indicators (U1, SST2 and NINO3) time series from 1939 to 2021.

ime series for streamflow energy for 4 subsystems in Brazil
}
\keyword{datasets}
