% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimaticInfo.R, R/StreamflowEnergy.R
\docType{data}
\name{ClimaticInfo}
\alias{ClimaticInfo}
\title{Climatic indicators for forecasting}
\format{
A four column data.frame with
\describe{
\item{Date is the Date for the observed values}
\item{U1 gives the Zonal Winds values}
\item{SST2 is the sea surface temperature in part of the Atlancic}
\item{NINO3 is the sea surface temperature at the Nino 3.4 region in the Pacific}
}

A five column data.frame with
\describe{
\item{Date is the Date for the observed values}
\item{Subsystem_N is the Northern Brazilian Subsystem}
\item{Subsystem_NE is the North East Brazilian Subsystem}
\item{Subsystem_S is the Southern Brazilian Subsystem}
\item{Subsystem_SE is the South East / Mid West Brazilian Subsystem}
}
}
\usage{
ClimaticInfo

ClimaticInfo
}
\description{
Three climatic indicators (U1, SST2 and NINO3) time series from 1939 to 2021.

ime series for streamflow energy for 4 subsystems in Brazil
}
\keyword{datasets}
